osd: Fix the OpSchedulerItem cost scaling calculation.
Calculate the scaled cost for an OpSchedulerItem in msec based on the
overall osd capacity (across all shards). Remove logic to cache the
cost of the previous op to re-use for the next transaction as it doesn't
seem to provide any benefit in terms of performance.

Signed-off-by: Sridhar Seshasayee <sseshasa@redhat.com>
(cherry picked from commit aebd9b71f29031dfb68964046af994bd89fea36b)