blk: avoid temporary bptrs on aio paths; use ptr_node instead.
This is a slight optimization for the `HAVE_LIBAIO` paths
of the kernel-based `BlockDevice` implementation.
The overall idea is to squeeze temporary, short-living
instances of `ceph::bufferptr` as `ceph::bufferlist`
actually aggregates `ptr_node` (`bufferptr` with the extra
`next` pointer field to form a list). It can be created
directly and this commit switches to exactly this behavior.

Signed-off-by: Radoslaw Zarzynski <rzarzyns@redhat.com>