rbd: add image map timeouts on Windows
This change adds configurable timeouts used when starting the centralized RBD
service and mapping images.

This change also fixes an issue where the service would wait indefinitely for
a failed mapping. This issue was caused by the fact that multiple child
processes were inheriting pipe handles.

Note that we can't use timeouts with Windows anonymous pipes, which is why
we're going to use unique named pipes.

Signed-off-by: Lucian Petrut <lpetrut@cloudbasesolutions.com>