mgr/volumes: deploy MDSs when creating fs
This line was (inadvertantly?) removed in 34f88e4585ba8029c45d7249347faaf63847d3c7
with the addition of the mds_autoscaler module.  However, that module
does not know how to deal with a filesystem that doesn't have a service
spec defined yet.  (Also, the module is not enabled by default.)

Signed-off-by: Sage Weil <sage@newdream.net>