mgr/volumes: deploy MDSs when creating fs
This line was (inadvertantly?) removed in f69abe6291958a9ccfb85ce77c9b86e8af8ecf21
with the addition of the mds_autoscaler module.  However, that module
does not know how to deal with a filesystem that doesn't have a service
spec defined yet.  (Also, the module is not enabled by default.)

Signed-off-by: Sage Weil <sage@newdream.net>