mgr/mds_autoscaler: some fixes and cleanup
Ignore filesystems that do not have a servicespec defined already, or
who have a service spec that does not include a count.  We can't scale
or manage these.

Ignore running daemons: we're just adjusting the service spec and
letting the orchestrator deploy those daemons accordingly.

Streamline log output a bit.

Signed-off-by: Sage Weil <sage@newdream.net>