mgr/cephadm:updates to support orch status changes
The available function definition has been updated in
cephadm/rook and Orchestrator base class to provide a
module specific dictionary holding any specific info that
would be pertinent to share with the user. This in turn
changed the _status method for cephadm, but the UX
remains the same (i.e. admin must use --detail to see
the worker pool size).

Signed-off-by: Paul Cuzner <pcuzner@redhat.com>