mgr: relax OSDs' ok-to-stop condition when the to-be-stopped OSDs holds only replicated pools
Right now, the "ok-to-stop" condition is relatively rigorous, it allows
stopping an osd only if no PG on it is non-active or degraded. But there
are situations in which an OSD is part of a degraded pg and the pg still
still have > min_size complete replicas after the OSD is stopped.

In [9750061](https://github.com/liewegas/ceph/commit/9750061d5d4236aaba156d60790e0b8bcd7cfb64),
we changed from considering just acting to using avail_no_missing (OSDs that have no missing
objects). When the projected pg_acting is constructed this way, we can safely compare to
min_size... even for a PG marked degraded.

Fixes: https://tracker.ceph.com/issues/49392
Signed-off-by: Xuehan Xu <xxhdx1985126@gmail.com>

Signed-off-by: Xuehan Xu <xxhdx1985126@gmail.com>