mgr/dashboard: improve rbd listing
Adds TTL caching to ioctx and RBD images in order to speed up listing of
many RBD images (> hundreds). It still does not fix the underlying
issue: since the number_of_rbd_images is a scaling factor, the latency
will sooner or later exceed the poll period.

The only fix for that is to introduce pagination. However, that will
break some Dashboard features, like the ability to sort RBD images by
any property (size, number of objects, provisioned size, etc.), or the
ability to filter those images by the same properties.

A possibility, already applied in other endpoints, is to disable the
periodic polling and allow the user to manually refresh the table.

Fixes: http://tracker.ceph.com/issues/39140
Signed-off-by: Ernesto Puerta <epuertat@redhat.com>