crimson/osd: do not pass ops to submit_transaction() separately
the order of evaluation of function arguments is unspecified with some
exceptions. but in this case, the reason why ops is still valid is that
the parameter type of `osd_op_p` is `const osd_op_params_t&`. so a copy
is passed to it.

to enable us to actually move `osd_op_p`, in this change, the
osd_op_p.req->ops is used instead.

Signed-off-by: Kefu Chai <kchai@redhat.com>