mon: convert ConfigKeyService -> KVMonitor
Convert this into a normal PaxosService.  This gives us (1) code uniformity
and also (2) a changelog in the form of the paxos commits for recently
changed keys that we can use to allow clients to subscribe to changes.

For upgrades this is pretty painless:
 - the actual kv data is in the same place
 - an old mon will still see updates made by a new monitor
 - a new mon will still see updates made by an old monitor, but won't get
   a paxosservice version bump.

Signed-off-by: Sage Weil <sage@newdream.net>
(cherry picked from commit bb7ebc41532aeb23cff2241ab07b3f01c2f57ddd)