mgr_module: get_foreign_ceph_option(entity, name)
Fetch a config value that applies to an arbitrary daemon in the cluster.

Initial implementaiton is not very efficient, but it is better than a
round-trip query to the monitor.

Note: this currently excludes mgr/ options (which should hopefully not be
of interest on other daemons!).

Signed-off-by: Sage Weil <sage@newdream.net>
(cherry picked from commit 02c44b1734d0c162acd77fb7989422591fbb3c8f)