mds: introduce ceph.mirror.dirty_snap_id vxattr
Purpose of this xattr is to identify if the "next" snapshot to be
synchronized can be incrementally transferred. Value of the xattr
is the snap-id of a snapshot in the primary cluster (for a given
directory root).

CephFS mirror daemon sets this xattr on the directory root once
the data for a snapshot is synchronized, thereby signifying that
the data under this directory root is the data of the snapshot
identified by the snap-id value. This allows the mirror daemon
to use the much efficient selective synchronzation (based on
mtime/ctime) by scanning the two snapshots locally and only
transferring the inodes that have changed.

Signed-off-by: Venky Shankar <vshankar@redhat.com>
(cherry picked from commit 544abaf720002bcfb02c1d3b325697542d0a501e)