mgr/cephadm:Multiple updates related to the addition of the CLI
Some changes needed to support the introduction of the CLI commands
used to manage the cephadm checks. For example, the main Cephadm
check class now interacts with the keystore directly to determine
status, and provides support for commands like ls to list the
check definitions. In addition the main class now handles existing
configuration checks and ensure that the stored state in the keystore
matches the checks defined by the module

Signed-off-by: Paul Cuzner <pcuzner@redhat.com>