client: wake up the front pos waiter
The old code was buggy, why it ran well without any problem is that
the big client_lock will make sure that the read/write won't run in
parallel, so it won't be any problem here.

But when introducing inode lock to break the client_lock, which will
allow the read/write to run parallelly, the pos waiters won't get any
chance to be woke up.

In unlock_fh_pos(), it will always wake up the frontest waiter.

Fixes: https://tracker.ceph.com/issues/49379
Signed-off-by: Xiubo Li <xiubli@redhat.com>