qa/crontab: adjust scheduled frequencies
smoke: 3x a week on master, weekly on other branches
pacific: every 2 days
master, octopus, nautilus, upgrades: weekly

This should be a net reduction in tests each week. Currently
we schedule far more than the lab can run in a given week.

Try to balance the scheduling over the course of the week.

Signed-off-by: Josh Durgin <jdurgin@redhat.com>