qa/suites/upgrade/octopus-x: do not test on centos (w/ podman 2.2.1)
We cannot upgrade *from* podman 2.2.1 because it fails to inspect the
newly pulled image.  Fixing that bug doesn't help because it is the old
version of the mgr that has to deploy a new mgr with new code.

For now, stop testing upgrades on 2.2.1 until this is fixed.

See https://tracker.ceph.com/issues/49522

Signed-off-by: Sage Weil <sage@newdream.net>