ceph.spec: switch over to ninja-build
for faster configure and build

the reason why we don't use cmake+ninja for building debian packages is
that the cmake+ninja support was added in debhelper v11.2.1. see
https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=895044

without which, we'd have to override override_dh_auto_build to use "ninja"
instead of "make", and specify the builddir explicitly everywhere.

it'd be better to wait until we ditch the bionic builds and do this to
save the efforts.

it's reported that OBS failed to build the packages with ninja, so
a "with ninja" rpm option is added so that downstream maintainer can
opt out this setting and continue using `make` as the CMake generator.

Signed-off-by: Kefu Chai <kchai@redhat.com>