osd: Add config options for cost per io & byte for the mclock scheduler
The cost per io and cost per byte options for hdd and ssd are specified
and set to default values determined using experiments on hdds and ssds
using a cost model. The values are used in calc_scaled_cost() to
determine the scaled cost for every OpSchedulerItem that is enqueued
within the mClockScheduler.

Signed-off-by: Sridhar Seshasayee <sseshasa@redhat.com>