mgr/DaemonServer: osd ok-to-stop: return json when there are unknown PGs
In 744ac52725f9a4641983c93afd5e4fc871c8e1e9 we switched to return JSON
both on success and fail to describe which PGs are affected or are blocking
the ability to stop/restart OSDs.  Do the same for the case where
some PG states are unknown (i.e., just after a mgr restart) so that
the cephadm upgrade process can unconditionally expect a JSON result.

Signed-off-by: Sage Weil <sage@newdream.net>