Revert "rgw: register daemon in service map with more details"
This reverts commit 46ec2f0ddd81b4b5b4a61633eada998dcd606c9b.

This broke cephadm (by triggering CEPHADM_STRAY_DAEMON) because cephadm
assumes that a daemon named rgw.r.z.foo will register as rgw.r.z.foo.

It is not clear to me that there is a way to work around this naming
mismatch that makes much sense.  I think it makes more sense to focus on
the use-case that needs daemons to register under unique names and perhaps
control that naming behavior via an option or invest in providing daemons
with unique ids up front.

Signed-off-by: Sage Weil <sage@newdream.net>