Revert "rgw: register daemon in service map with more details"
This reverts commit c51e2d0bfc48621f9b4a97c19f9d5b89e497d92a.

This broke cephadm (by triggering CEPHADM_STRAY_DAEMON) because cephadm
assumes that a daemon named rgw.r.z.foo will register as rgw.r.z.foo.

It is not clear to me that there is a way to work around this naming
mismatch that makes much sense.  I think it makes more sense to focus on
the use-case that needs daemons to register under unique names and perhaps
control that naming behavior via an option or invest in providing daemons
with unique ids up front.

Signed-off-by: Sage Weil <sage@newdream.net>