common/lockdep: introduce MAX_FOLLOWERS
before this change we use a matrix for tracking the locks and those
which are locked after acquiring the formers. where follows[a][b]
sigifies that b is acquired after a is acquired, and a and b are both
lock ids allocated in the lock_ids registry.

but since we might need to have a much large set of locks which are
acquired at the same time. but the numbers of their follower of
each of them are always much smaller. because we are not likely
to hold, for instance, 4096 locks when we still holding a lock.
actually, this might be a bug, if we actually do this.

so we can use smaller vectors for tracking the followers. and the
size of those vectors can be one or more orders of magnitude smaller
than the number of locks.

in this change, MAX_FOLLOWERS in introduced for the max number
of followers.

Signed-off-by: Kefu Chai <kchai@redhat.com>