rgw:fix the bug of file decompression error which is larger than 15M.
Last patch of compressor_message has a bug released. This patch is to fix the bug of file decompression, when the file is larger than 15M.
I see the last patch of compressor_message is merged in v16.1.0. As this patch is based on the last one, so it should be merged after the last patch.

Signed-off-by: zhouyunkai <498549175@qq.com>