rgw: add option to uniquely register in servicemap
If rgw_register_with_unique_name is set to True, the rgw process will
register in the service map with a uniq identifier. The uniq identifier
is the RADOS client session ID appended to the name of the gateway.

This will give us better visibility when
query the Ceph status and also will also us to run multiple rgws with
the same cephX user AND to be displayed individually. Previously, using a
single cephX user to run multiple gateways was possible but the Ceph
status would display a single rgw.
Also, in the same situation it will allow us to gather metrics and
export them through the ceph-mgr Prometheus exporter.

Fixes: https://tracker.ceph.com/issues/49227
Signed-off-by: Sébastien Han <seb@redhat.com>