lockdep: switch follows_bt vector member to std::map
Delete the MAX_FOLLOWERS and increase the bitset size to MAX_LOCKS
in 'follows'. The maximum memories of the 'follows' will be 2GB.

But for the 'follows_bt', if we continue use the std::array, the
maximum memories will be 128G, but most of the array spaces are
useless. Switch the std::array to std::man instead.

Signed-off-by: Xiubo Li <xiubli@redhat.com>