mgr/dashboard: provide the service events when showing a service in the UI
When service deployment failures occur, events are generated and associated with the service. This PR intends to add these events as Daemon Logs and Service Logs to the UI for troubleshooting and doing diagnostics for the same.

Fixes: https://tracker.ceph.com/issues/49262
Signed-off-by: Aashish Sharma <aasharma@redhat.com>
(cherry picked from commit 00ce7c90ef19f02beea9431ab360d23744efde2d)