qa/tasks/mgr/test_progress.py: remove _osd_in_out_completed_events_count()
delete the part where  _osd_in_out_completed_events_count()
was called in test_osd_cannot_recover() and revert to initial
state of the function since we don't need to use this function
in octopus. Also delete a duplicate of _osd_in_out_events_count().
This must be added by mistake in #39289 as well.

No need to fix for the backport in Nautilus: #38173
since the bugs are occured by adding additional code to
the cherry-pick specifically for Octopus.

fixes: https://tracker.ceph.com/issues/49891

Signed-off-by: Kamoltat <ksirivad@redhat.com>